requirements.txt
alabaster==0.7.12
asgiref==3.5.0
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
Django==3.2.13
django-cors-headers==3.13.0
djangorestframework==3.13.1
docutils==0.17.1
fonttools==4.33.3
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.21.6
packaging==21.3
pandas==1.5.0
Pillow==9.1.0
psycopg2==2.9.3
Pygments==2.13.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scipy==1.7.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.11
